@@ -79,33 +79,35 @@
     "test:types": "tsc -p ./types"
   },
   "dependencies": {
-    "sax": "^1.2.4"
+    "sax": "^1.4.3"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.28.5",
     "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
+    "babel-loader": "^10.0.0",
     "babel-preset-env": "^1.7.0",
     "biased-opener": "^0.2.8",
-    "browser-sync": "^2.26.3",
+    "browser-sync": "^3.0.4",
     "cash-cat": "^0.2.0",
     "codacy-coverage": "^3.4.0",
     "codeclimate-test-reporter": "^0.5.1",
-    "codecov": "^3.1.0",
-    "coveralls": "^3.0.2",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.12.0",
-    "globify": "^2.3.4",
+    "codecov": "^3.8.3",
+    "coveralls": "^3.1.1",
+    "cross-env": "^10.1.0",
+    "eslint": "^9.39.2",
+    "globify": "^3.0.3",
     "istanbul": "^0.4.5",
-    "jasmine": "^3.3.1",
-    "jest": "^20.0.4",
-    "jest-cli": "^20.0.4",
-    "jsonpath": "^1.0.0",
-    "nodemon": "^1.18.9",
+    "jasmine": "^5.13.0",
+    "jest": "^30.2.0",
+    "jest-cli": "^30.2.0",
+    "jest-html-reporter": "^4.3.0",
+    "jsonpath": "^1.1.1",
+    "nodemon": "^3.1.11",
     "npm-run-all": "^4.1.5",
-    "prismjs": "^1.15.0",
-    "typescript": "^3.2.2",
-    "unminified-webpack-plugin": "^1.4.2",
-    "watch": "^1.0.1",
-    "webpack": "^3.10.0"
+    "prismjs": "^1.30.0",
+    "typescript": "^5.9.3",
+    "unminified-webpack-plugin": "^3.0.0",
+    "watch": "^1.0.2",
+    "webpack": "^5.104.1"
   }
 }
