@@ -1,49 +1,48 @@
-var helper = require('./options-helper');
-var isArray = require('./array-helper').isArray;
+import { copyOptions, ensureFlagExists, ensureSpacesExists, ensureKeyExists, checkFnExists } from './options-helper';
 
 var currentElement, currentElementName;
 
 function validateOptions(userOptions) {
-  var options = helper.copyOptions(userOptions);
-  helper.ensureFlagExists('ignoreDeclaration', options);
-  helper.ensureFlagExists('ignoreInstruction', options);
-  helper.ensureFlagExists('ignoreAttributes', options);
-  helper.ensureFlagExists('ignoreText', options);
-  helper.ensureFlagExists('ignoreComment', options);
-  helper.ensureFlagExists('ignoreCdata', options);
-  helper.ensureFlagExists('ignoreDoctype', options);
-  helper.ensureFlagExists('compact', options);
-  helper.ensureFlagExists('indentText', options);
-  helper.ensureFlagExists('indentCdata', options);
-  helper.ensureFlagExists('indentAttributes', options);
-  helper.ensureFlagExists('indentInstruction', options);
-  helper.ensureFlagExists('fullTagEmptyElement', options);
-  helper.ensureFlagExists('noQuotesForNativeAttributes', options);
-  helper.ensureSpacesExists(options);
+  var options = copyOptions(userOptions);
+  ensureFlagExists('ignoreDeclaration', options);
+  ensureFlagExists('ignoreInstruction', options);
+  ensureFlagExists('ignoreAttributes', options);
+  ensureFlagExists('ignoreText', options);
+  ensureFlagExists('ignoreComment', options);
+  ensureFlagExists('ignoreCdata', options);
+  ensureFlagExists('ignoreDoctype', options);
+  ensureFlagExists('compact', options);
+  ensureFlagExists('indentText', options);
+  ensureFlagExists('indentCdata', options);
+  ensureFlagExists('indentAttributes', options);
+  ensureFlagExists('indentInstruction', options);
+  ensureFlagExists('fullTagEmptyElement', options);
+  ensureFlagExists('noQuotesForNativeAttributes', options);
+  ensureSpacesExists(options);
   if (typeof options.spaces === 'number') {
     options.spaces = Array(options.spaces + 1).join(' ');
   }
-  helper.ensureKeyExists('declaration', options);
-  helper.ensureKeyExists('instruction', options);
-  helper.ensureKeyExists('attributes', options);
-  helper.ensureKeyExists('text', options);
-  helper.ensureKeyExists('comment', options);
-  helper.ensureKeyExists('cdata', options);
-  helper.ensureKeyExists('doctype', options);
-  helper.ensureKeyExists('type', options);
-  helper.ensureKeyExists('name', options);
-  helper.ensureKeyExists('elements', options);
-  helper.checkFnExists('doctype', options);
-  helper.checkFnExists('instruction', options);
-  helper.checkFnExists('cdata', options);
-  helper.checkFnExists('comment', options);
-  helper.checkFnExists('text', options);
-  helper.checkFnExists('instructionName', options);
-  helper.checkFnExists('elementName', options);
-  helper.checkFnExists('attributeName', options);
-  helper.checkFnExists('attributeValue', options);
-  helper.checkFnExists('attributes', options);
-  helper.checkFnExists('fullTagEmptyElement', options);
+  ensureKeyExists('declaration', options);
+  ensureKeyExists('instruction', options);
+  ensureKeyExists('attributes', options);
+  ensureKeyExists('text', options);
+  ensureKeyExists('comment', options);
+  ensureKeyExists('cdata', options);
+  ensureKeyExists('doctype', options);
+  ensureKeyExists('type', options);
+  ensureKeyExists('name', options);
+  ensureKeyExists('elements', options);
+  checkFnExists('doctype', options);
+  checkFnExists('instruction', options);
+  checkFnExists('cdata', options);
+  checkFnExists('comment', options);
+  checkFnExists('text', options);
+  checkFnExists('instructionName', options);
+  checkFnExists('elementName', options);
+  checkFnExists('attributeName', options);
+  checkFnExists('attributeValue', options);
+  checkFnExists('attributes', options);
+  checkFnExists('fullTagEmptyElement', options);
   return options;
 }
 
@@ -282,7 +281,7 @@ function writeElementsCompact(element, options, depth, firstLine) {
   var i, key, nodes, xml = [];
   for (key in element) {
     if (element.hasOwnProperty(key)) {
-      nodes = isArray(element[key]) ? element[key] : [element[key]];
+      nodes = Array.isArray(element[key]) ? element[key] : [element[key]];
       for (i = 0; i < nodes.length; ++i) {
         switch (key) {
         case options.declarationKey: xml.push(writeDeclaration(nodes[i], options, depth)); break;
@@ -301,7 +300,7 @@ function writeElementsCompact(element, options, depth, firstLine) {
   return xml.join('');
 }
 
-module.exports = function (js, options) {
+export default function (js, options) {
   options = validateOptions(options);
   var xml = [];
   currentElement = js;
