@@ -1,6 +1,6 @@
-var js2xml = require('./js2xml.js');
+import js2xml from './js2xml.js';
 
-module.exports = function (json, options) {
+export default function (json, options) {
   if (json instanceof Buffer) {
     json = json.toString();
   }
