@@ -1,52 +1,51 @@
-var sax = require('sax');
+import { parser as _parser } from 'sax';
 var expat /*= require('node-expat');*/ = { on: function () { }, parse: function () { } };
-var helper = require('./options-helper');
-var isArray = require('./array-helper').isArray;
+import { copyOptions, ensureFlagExists, ensureAlwaysArrayExists, ensureKeyExists, checkFnExists } from './options-helper';
 
 var options;
 var pureJsParser = true;
 var currentElement;
 
 function validateOptions(userOptions) {
-  options = helper.copyOptions(userOptions);
-  helper.ensureFlagExists('ignoreDeclaration', options);
-  helper.ensureFlagExists('ignoreInstruction', options);
-  helper.ensureFlagExists('ignoreAttributes', options);
-  helper.ensureFlagExists('ignoreText', options);
-  helper.ensureFlagExists('ignoreComment', options);
-  helper.ensureFlagExists('ignoreCdata', options);
-  helper.ensureFlagExists('ignoreDoctype', options);
-  helper.ensureFlagExists('compact', options);
-  helper.ensureFlagExists('alwaysChildren', options);
-  helper.ensureFlagExists('addParent', options);
-  helper.ensureFlagExists('trim', options);
-  helper.ensureFlagExists('nativeType', options);
-  helper.ensureFlagExists('nativeTypeAttributes', options);
-  helper.ensureFlagExists('sanitize', options);
-  helper.ensureFlagExists('instructionHasAttributes', options);
-  helper.ensureFlagExists('captureSpacesBetweenElements', options);
-  helper.ensureAlwaysArrayExists(options);
-  helper.ensureKeyExists('declaration', options);
-  helper.ensureKeyExists('instruction', options);
-  helper.ensureKeyExists('attributes', options);
-  helper.ensureKeyExists('text', options);
-  helper.ensureKeyExists('comment', options);
-  helper.ensureKeyExists('cdata', options);
-  helper.ensureKeyExists('doctype', options);
-  helper.ensureKeyExists('type', options);
-  helper.ensureKeyExists('name', options);
-  helper.ensureKeyExists('elements', options);
-  helper.ensureKeyExists('parent', options);
-  helper.checkFnExists('doctype', options);
-  helper.checkFnExists('instruction', options);
-  helper.checkFnExists('cdata', options);
-  helper.checkFnExists('comment', options);
-  helper.checkFnExists('text', options);
-  helper.checkFnExists('instructionName', options);
-  helper.checkFnExists('elementName', options);
-  helper.checkFnExists('attributeName', options);
-  helper.checkFnExists('attributeValue', options);
-  helper.checkFnExists('attributes', options);
+  options = copyOptions(userOptions);
+  ensureFlagExists('ignoreDeclaration', options);
+  ensureFlagExists('ignoreInstruction', options);
+  ensureFlagExists('ignoreAttributes', options);
+  ensureFlagExists('ignoreText', options);
+  ensureFlagExists('ignoreComment', options);
+  ensureFlagExists('ignoreCdata', options);
+  ensureFlagExists('ignoreDoctype', options);
+  ensureFlagExists('compact', options);
+  ensureFlagExists('alwaysChildren', options);
+  ensureFlagExists('addParent', options);
+  ensureFlagExists('trim', options);
+  ensureFlagExists('nativeType', options);
+  ensureFlagExists('nativeTypeAttributes', options);
+  ensureFlagExists('sanitize', options);
+  ensureFlagExists('instructionHasAttributes', options);
+  ensureFlagExists('captureSpacesBetweenElements', options);
+  ensureAlwaysArrayExists(options);
+  ensureKeyExists('declaration', options);
+  ensureKeyExists('instruction', options);
+  ensureKeyExists('attributes', options);
+  ensureKeyExists('text', options);
+  ensureKeyExists('comment', options);
+  ensureKeyExists('cdata', options);
+  ensureKeyExists('doctype', options);
+  ensureKeyExists('type', options);
+  ensureKeyExists('name', options);
+  ensureKeyExists('elements', options);
+  ensureKeyExists('parent', options);
+  checkFnExists('doctype', options);
+  checkFnExists('instruction', options);
+  checkFnExists('cdata', options);
+  checkFnExists('comment', options);
+  checkFnExists('text', options);
+  checkFnExists('instructionName', options);
+  checkFnExists('elementName', options);
+  checkFnExists('attributeName', options);
+  checkFnExists('attributeValue', options);
+  checkFnExists('attributes', options);
   return options;
 }
 
@@ -69,11 +68,11 @@ function addField(type, value) {
   if (options.compact) {
     if (
       !currentElement[options[type + 'Key']] &&
-      (isArray(options.alwaysArray) ? options.alwaysArray.indexOf(options[type + 'Key']) !== -1 : options.alwaysArray)
+      (Array.isArray(options.alwaysArray) ? options.alwaysArray.indexOf(options[type + 'Key']) !== -1 : options.alwaysArray)
     ) {
       currentElement[options[type + 'Key']] = [];
     }
-    if (currentElement[options[type + 'Key']] && !isArray(currentElement[options[type + 'Key']])) {
+    if (currentElement[options[type + 'Key']] && !Array.isArray(currentElement[options[type + 'Key']])) {
       currentElement[options[type + 'Key']] = [currentElement[options[type + 'Key']]];
     }
     if (type + 'Fn' in options && typeof value === 'string') {
@@ -92,7 +91,7 @@ function addField(type, value) {
         }
       }
     }
-    if (isArray(currentElement[options[type + 'Key']])) {
+    if (Array.isArray(currentElement[options[type + 'Key']])) {
       currentElement[options[type + 'Key']].push(value);
     } else {
       currentElement[options[type + 'Key']] = value;
@@ -220,14 +219,14 @@ function onStartElement(name, attributes) {
     }
     if (
       !(name in currentElement) &&
-      (isArray(options.alwaysArray) ? options.alwaysArray.indexOf(name) !== -1 : options.alwaysArray)
+      (Array.isArray(options.alwaysArray) ? options.alwaysArray.indexOf(name) !== -1 : options.alwaysArray)
     ) {
       currentElement[name] = [];
     }
-    if (currentElement[name] && !isArray(currentElement[name])) {
+    if (currentElement[name] && !Array.isArray(currentElement[name])) {
       currentElement[name] = [currentElement[name]];
     }
-    if (isArray(currentElement[name])) {
+    if (Array.isArray(currentElement[name])) {
       currentElement[name].push(element);
     } else {
       currentElement[name] = element;
@@ -313,9 +312,9 @@ function onError(error) {
   error.note = error; //console.error(error);
 }
 
-module.exports = function (xml, userOptions) {
+export default function (xml, userOptions) {
 
-  var parser = pureJsParser ? sax.parser(true, {}) : parser = new expat.Parser('UTF-8');
+  var parser = pureJsParser ? _parser(true, {}) : parser = new expat.Parser('UTF-8');
   var result = {};
   currentElement = result;
 
