@@ -1,13 +1,13 @@
-var helper = require('./options-helper');
-var xml2js = require('./xml2js');
+import { copyOptions, ensureSpacesExists } from './options-helper';
+import xml2js from './xml2js';
 
 function validateOptions (userOptions) {
-  var options = helper.copyOptions(userOptions);
-  helper.ensureSpacesExists(options);
+  var options = copyOptions(userOptions);
+  ensureSpacesExists(options);
   return options;
 }
 
-module.exports = function(xml, userOptions) {
+export default function(xml, userOptions) {
   var options, js, json, parentKey;
   options = validateOptions(userOptions);
   js = xml2js(xml, options);
