@@ -1,9 +1,9 @@
-var path = require('path');
+import { join, relative } from 'path';
 var jasmine = new (require('jasmine'))();
 var config = require(__dirname + '/jasmine.json');
 
 if ('spec_dir' in config) {
-  config.spec_dir = path.join(path.relative(process.cwd(), __dirname), '..', config.spec_dir);
+  config.spec_dir = join(relative(process.cwd(), __dirname), '..', config.spec_dir);
 }
 
 jasmine.loadConfig(config);
