@@ -1,34 +1,29 @@
-const path = require('path');
+import { fileURLToPath } from "url";
+import { dirname } from "path";
 
-module.exports = {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
-  rootDir: path.dirname(__dirname),
-
-  modulePaths: [
-    '<rootDir>/lib',
-    '<rootDir>/node_modules'
+export const rootDir = dirname(__dirname);
+export const modulePaths = ["<rootDir>/lib", "<rootDir>/node_modules"];
+export const reporters = [
+  "default",
+  [
+    "./node_modules/jest-html-reporter",
+    {
+      pageTitle: "Test Report",
+    },
   ],
-
-  moduleFileExtensions: ['js'],
-
-  // setupFiles: ['<rootDir>/test/pretest-jest.js'], // don't use this, expect.extend() will fail
-  setupTestFrameworkScriptFile: '<rootDir>/test/pretest-jest.js',
-
-  testEnvironment: 'node',
-
-  testMatch: ['**/*.spec.(ts|js)?(x)'],
-
-  collectCoverage: true,
-
-  collectCoverageFrom: [
-    'src-js/**/*.{js,ts}',
-    '!**/*.spec.{js,ts}',
-    '!**/node_modules/**',
-    '!**/test/**'
-  ],
-
-  coverageDirectory: '<rootDir>/test/coverage-jest',
-
-  coverageReporters: ['json', 'lcov', 'text', 'html']
-
-};
+];
+export const moduleFileExtensions = ["js"];
+export const testEnvironment = "node";
+export const testMatch = ["**/*.spec.(ts|js)?(x)"];
+export const collectCoverage = true;
+export const collectCoverageFrom = [
+  "src-js/**/*.{js,ts}",
+  "!**/*.spec.{js,ts}",
+  "!**/node_modules/**",
+  "!**/test/**",
+];
+export const coverageDirectory = "<rootDir>/test/coverage-jest";
+export const coverageReporters = ["json", "lcov", "text", "html"];
