@@ -1,4 +1,4 @@
-var convert = require('..');
+import { xml2json } from '..';
 var xml =
 '<?xml version="1.0" encoding="utf-8"?>' + '\n' +
 '<note importance="high" logged="true">' + '\n' +
@@ -6,6 +6,6 @@ var xml =
 '    <todo>Work</todo>' + '\n' +
 '    <todo>Play</todo>' + '\n' +
 '</note>';
-var result1 = convert.xml2json(xml, {compact: true});
-var result2 = convert.xml2json(xml, {compact: false});
+var result1 = xml2json(xml, {compact: true});
+var result2 = xml2json(xml, {compact: false});
 console.log(result1, '\n', result2);
