@@ -1,5 +1,5 @@
-var convert = require('../lib');
-var testItems = require('./test-items');
+import { xml2js } from '../lib';
+import testItems from './test-items';
 
 /*global describe,it,expect*/
 
@@ -12,7 +12,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, declarationKey: 'declaration'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -23,7 +23,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, instructionKey: 'instruction'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -34,7 +34,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, attributesKey: 'attributes'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -45,7 +45,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, textKey: 'text'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -56,7 +56,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, cdataKey: 'cdata'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -67,7 +67,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, doctypeKey: 'doctype'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -78,7 +78,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, commentKey: 'comment'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -89,7 +89,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, parentKey: 'parent'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -100,7 +100,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, typeKey: 'type'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -111,7 +111,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, nameKey: 'name'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -122,7 +122,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: false, elementsKey: 'elements'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -137,7 +137,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, declarationKey: 'declaration'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -148,7 +148,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, instructionKey: 'instruction'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -159,7 +159,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, attributesKey: 'attributes'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -170,7 +170,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, textKey: 'text'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -181,7 +181,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, cdataKey: 'cdata'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -192,7 +192,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, doctypeKey: 'doctype'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -203,7 +203,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, commentKey: 'comment'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -214,7 +214,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, parentKey: 'parent'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -225,7 +225,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, typeKey: 'type'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -236,7 +236,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, nameKey: 'name'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
@@ -247,7 +247,7 @@ describe('Testing xml2js.js:', function () {
       var options = {compact: true, elementsKey: 'elements'.slice(0,3)};
       testItems('xml2js', options).forEach(function (test) {
         it(test.desc, function () {
-          expect(convert.xml2js(test.xml, options)).toEqual(test.js);
+          expect(xml2js(test.xml, options)).toEqual(test.js);
         });
       });
 
