@@ -1,13 +1,6 @@
 /*jslint node:true */
 
-var xml2js = require('./xml2js');
-var xml2json = require('./xml2json');
-var js2xml = require('./js2xml');
-var json2xml = require('./json2xml');
-
-module.exports = {
-  xml2js: xml2js,
-  xml2json: xml2json,
-  js2xml: js2xml,
-  json2xml: json2xml
-};
+export { default as xml2js } from "./xml2js";
+export { default as xml2json } from "./xml2json";
+export { default as js2xml } from "./js2xml";
+export { default as json2xml } from "./json2xml";
