@@ -1,3 +1,3 @@
 expect.extend({
-  toContain: (collection, value) => require('jest-matchers/build/matchers.js').toContainEqual(collection, value)
+  toContain: (collection, value) => jest.toContainEqual(collection, value)
 });
